fix error install on postgresql
